+++ Tests/Objects/IGTestDelegateController.h
@@ -17,6 +17,8 @@
 
 @property (nonatomic, strong, readonly) IGTestObject *item;
 
+@property (nonatomic, assign) CGFloat height;
+
 @property (nonatomic, copy) void (^itemUpdateBlock)();
 @property (nonatomic, copy) void (^cellConfigureBlock)(IGTestDelegateController *);
 @property (nonatomic, assign, readonly) NSInteger updateCount;